// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ServiceStatus = {
  ativo: 'ativo',
  inativo: 'inativo'
} as const

export type ServiceStatus = (typeof ServiceStatus)[keyof typeof ServiceStatus]


export const OrderStatus = {
  PENDENTE: 'PENDENTE',
  EM_ANDAMENTO: 'EM_ANDAMENTO',
  CONCLUIDO: 'CONCLUIDO',
  CANCELADO: 'CANCELADO'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const AppointmentStatus = {
  CONFIRMADO: 'CONFIRMADO',
  PENDENTE: 'PENDENTE',
  CONCLUIDO: 'CONCLUIDO',
  CANCELADO: 'CANCELADO'
} as const

export type AppointmentStatus = (typeof AppointmentStatus)[keyof typeof AppointmentStatus]